Change: remove release creation; use artifact instead
Using github releases is not really good for version management, as this is more of "nightly" build. Changing to artifacts is probably more flexible; on top of that, this avoids the tagging problem.

So the question here comes: WHY THE HELL would I choose to use to create releases and pre-releases?